% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cache_records.R
\name{clean_cache_records}
\alias{clean_cache_records}
\title{clean up cache records}
\usage{
clean_cache_records(cache_dir = ".")
}
\arguments{
\item{cache_dir}{cache folder location}
}
\description{
clean up cache records
}
\details{
cleans up the cache record to only include files that currently exist in the cache, useful to run
if manually deleting files in the cache
}

