% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-remote-xpt.R
\name{read_remote_factory}
\alias{read_remote_factory}
\title{create remote reader functions}
\usage{
read_remote_factory(cb)
}
\arguments{
\item{cb}{callback function that is applied to read the remote data}
}
\description{
create remote reader functions
}
\examples{
\dontrun{
read_remote_csv <- read_remote_factory(readr::read_csv)
data <- read_remote_csv("<remote_dir/remote_file.csv>", "~/.cache")
}

}

