% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-remote-xpt.R
\name{read_remote_xpt}
\alias{read_remote_csv}
\alias{read_remote_sas}
\alias{read_remote_xlsx}
\alias{read_remote_xpt}
\title{read remote files}
\usage{
read_remote_xpt(remote, cache_dir, ...)

read_remote_csv(remote, cache_dir, ...)

read_remote_xlsx(remote, cache_dir, ...)

read_remote_sas(remote, cache_dir, ...)
}
\arguments{
\item{remote}{remote data file}

\item{cache_dir}{local cache folder}

\item{...}{args to pass to read function}
}
\description{
read remote files
}
\examples{
\dontrun{
example_data <- read_remote_xpt("\\\\cdsnas\\PHARMACOMETRICS\\Fellows\\Devin\\example.xpt", ".cache")
}
}

