% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-remote-xpt.R
\name{read_remote_xptc}
\alias{read_remote_csvc}
\alias{read_remote_sasc}
\alias{read_remote_xlsxc}
\alias{read_remote_xptc}
\title{read remote files with compression}
\usage{
read_remote_xptc(remote, cache_dir, ...)

read_remote_csvc(remote, cache_dir, ...)

read_remote_xlsxc(remote, cache_dir, ...)

read_remote_sasc(remote, cache_dir, ...)
}
\arguments{
\item{remote}{remote data file}

\item{cache_dir}{local cache folder}

\item{...}{args to pass to read function}
}
\description{
read remote files with compression
}

