% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-remote-xpt.R
\name{read_remote_factory}
\alias{read_remote_factory}
\title{create remote reader functions}
\usage{
read_remote_factory(cb, cache_method = "rds", compression = FALSE)
}
\arguments{
\item{cb}{callback function that is applied to read the remote data}

\item{cache_method}{caching method, either rds or feather}

\item{compression}{whether to compress files if saving as rds, default to false}
}
\description{
create remote reader functions
}
\examples{
\dontrun{
read_remote_csv <- read_remote_factory(readr::read_csv)
data <- read_remote_csv("<remote_dir/remote_file.csv>", "~/.cache")
}

}

